refactor(pages/home): extract `<Stats />`
